+++ Source/DataSources/BillboardVisualizer.js
@@ -80,7 +80,7 @@ define([
         this._billboardCollection = undefined;
         this._entityCollection = entityCollection;
         this._items = new AssociativeArray();
-        this._onCollectionChanged(entityCollection, entityCollection.entities, [], []);
+        this._onCollectionChanged(entityCollection, entityCollection.values, [], []);
     };
 
     /**