+++ Source/DataSources/EntityCollection.js
@@ -4,6 +4,7 @@ define([
         '../Core/createGuid',
         '../Core/defined',
         '../Core/defineProperties',
+        '../Core/deprecationWarning',
         '../Core/DeveloperError',
         '../Core/Event',
         '../Core/Iso8601',
@@ -16,6 +17,7 @@ define([
         createGuid,
         defined,
         defineProperties,
+        deprecationWarning,
         DeveloperError,
         Event,
         Iso8601,
@@ -131,8 +133,22 @@ define([
          * @memberof EntityCollection.prototype
          * @readonly
          * @type {Entity[]}
+         * @deprecated
          */
         entities : {
+            get : function() {
+                deprecationWarning('EntityCollection.entities', 'EntityCollection.entities has been deprecated and will be removed in Cesium 1.9, use EntityCollection.values instead');
+                return this._entities.values;
+            }
+        },
+        /**
+         * Gets the array of Entity instances in the collection.
+         * This array should not be modified directly.
+         * @memberof EntityCollection.prototype
+         * @readonly
+         * @type {Entity[]}
+         */
+        values : {
             get : function() {
                 return this._entities.values;
             }