+++ Source/DataSources/PointVisualizer.js
@@ -67,7 +67,7 @@ define([
         this._entityCollection = entityCollection;
         this._billboardCollection = undefined;
         this._items = new AssociativeArray();
-        this._onCollectionChanged(entityCollection, entityCollection.entities, [], []);
+        this._onCollectionChanged(entityCollection, entityCollection.values, [], []);
     };
 
     /**